Moved SqlMode to separate file for Antlr4ng. Fixed constructor for lexer and parser in TypeScript port.
