+++ sql/mysql/Oracle/TypeScript/MySQLParserBase.ts
@@ -2,7 +2,7 @@
  * Copyright © 2024, Oracle and/or its affiliates
  */
 
-import { Parser } from "antlr4";
+import { Parser, TokenStream } from "antlr4";
 import MySQLParser from "./MySQLParser";
 import SqlMode from "./SqlMode";
 
@@ -14,6 +14,12 @@ export default abstract class MySQLParserBase extends Parser {
 
     /** Enable Multi Language Extension support. */
     public supportMle = true;
+
+    constructor(input: TokenStream) {
+        super(input);
+        this.serverVersion = 80200;
+        this.sqlModeFromString("ANSI_QUOTES");
+    }
 
     /**
      * Determines if the given SQL mode is currently active in the lexer.
@@ -28,21 +34,48 @@ export default abstract class MySQLParserBase extends Parser {
 
     public isPureIdentifier(): boolean
     {
-	return this.isSqlModeActive(SqlMode.AnsiQuotes);
+        return this.isSqlModeActive(SqlMode.AnsiQuotes);
     }
 
     public isTextStringLiteral(): boolean
     {
-	return !this.isSqlModeActive(SqlMode.AnsiQuotes);
+        return !this.isSqlModeActive(SqlMode.AnsiQuotes);
     }
 
     public isStoredRoutineBody(): boolean
     {
-	return this.serverVersion >= 80032 && this.supportMle;
+        return this.serverVersion >= 80032 && this.supportMle;
     }
 
     public isSelectStatementWithInto(): boolean
     {
-	return this.serverVersion >= 80024 && this.serverVersion < 80031;
+        return this.serverVersion >= 80024 && this.serverVersion < 80031;
+    }
+
+    /**
+     * Converts a mode string into individual mode flags.
+     *
+     * @param modes The input string to parse.
+     */
+    public sqlModeFromString(modes: string): void {
+        this.sqlModes = new Set<SqlMode>();
+
+        const parts = modes.toUpperCase().split(",");
+        parts.forEach((mode: string) => {
+            if (mode === "ANSI" || mode === "DB2" || mode === "MAXDB" || mode === "MSSQL" || mode === "ORACLE" ||
+                mode === "POSTGRESQL") {
+                this.sqlModes.add(SqlMode.AnsiQuotes).add(SqlMode.PipesAsConcat).add(SqlMode.IgnoreSpace);
+            } else if (mode === "ANSI_QUOTES") {
+                this.sqlModes.add(SqlMode.AnsiQuotes);
+            } else if (mode === "PIPES_AS_CONCAT") {
+                this.sqlModes.add(SqlMode.PipesAsConcat);
+            } else if (mode === "NO_BACKSLASH_ESCAPES") {
+                this.sqlModes.add(SqlMode.NoBackslashEscapes);
+            } else if (mode === "IGNORE_SPACE") {
+                this.sqlModes.add(SqlMode.IgnoreSpace);
+            } else if (mode === "HIGH_NOT_PRECEDENCE" || mode === "MYSQL323" || mode === "MYSQL40") {
+                this.sqlModes.add(SqlMode.HighNotPrecedence);
+            }
+        });
     }
 }
